refactor : 로그아웃 api에 UserContext 활용
