+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/infrastructure/RefreshTokenRepository.kt
@@ -5,6 +5,6 @@ import org.depromeet.clog.server.domain.user.domain.Provider
 
 interface RefreshTokenRepository {
     fun save(token: RefreshToken): RefreshToken
-    fun findById(id: String): RefreshToken?
+    fun findByUserId(id: Long): RefreshToken?
     fun deleteByUserIdAndProvider(userId: Long, provider: Provider)
 }
