+++ clog-infrastructure/src/main/kotlin/org/depromeet/clog/server/infrastructure/auth/RefreshTokenRepositoryAdapter.kt
@@ -3,7 +3,6 @@ package org.depromeet.clog.server.infrastructure.auth
 import org.depromeet.clog.server.domain.auth.domain.RefreshToken
 import org.depromeet.clog.server.domain.auth.infrastructure.RefreshTokenRepository
 import org.depromeet.clog.server.domain.user.domain.Provider
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 
 @Component
@@ -16,8 +15,8 @@ class RefreshTokenRepositoryAdapter(
         return refreshTokenJpaRepository.save(entity).toDomain()
     }
 
-    override fun findById(id: String): RefreshToken? =
-        refreshTokenJpaRepository.findByIdOrNull(id)?.toDomain()
+    override fun findByUserId(id: Long): RefreshToken? =
+        refreshTokenJpaRepository.findByUserId(id)?.toDomain()
 
     override fun deleteByUserIdAndProvider(userId: Long, provider: Provider) =
         refreshTokenJpaRepository.deleteByUserIdAndProvider(userId, provider)
